requirements.txt
python-telegram-bot==21.10
groq==0.25.0
python-dotenv==1.1.0
